anti3/assistant/requirements.txt
# Hardware
adafruit-circuitpython-ssd1306
adafruit-blinka
gpiod

# AI / Vision
ultralytics
llama-cpp-python

# Audio
sounddevice
scipy
faster-whisper
piper-tts

# Core
Pillow
numpy
